+++ module2/4.3/main.c
@@ -164,7 +164,7 @@ void print_address(const address *a)
 
 void print_contact(const contact *c)
 {
-    printf("Контакт #%llu:\n", c->id);
+    printf("Контакт #%lu:\n", c->id);
 
     print_names(&c->names);
     print_address(&c->address);
@@ -323,11 +323,11 @@ void print_tree_recursive(node *node, char *prefix, int is_left, int is_root)
 
     if (is_root)
     {
-        printf("%llu\n", ((contact *)node->value)->id);
+        printf("%lu\n", ((contact *)node->value)->id);
     }
     else
     {
-        printf("%s%s── %llu\n", prefix, is_left ? "├" : "└", ((contact *)node->value)->id);
+        printf("%s%s── %lu\n", prefix, is_left ? "├" : "└", ((contact *)node->value)->id);
     }
     char *new_prefix = malloc(strlen(prefix) + 7);
 
@@ -398,8 +398,8 @@ void menu2(phonebook *pb)
     }
 
     printf("Введите id контакта для удаления: ");
-    unsigned long long id;
-    scanf("%llu", &id);
+    uint64_t id;
+    scanf("%lu", &id);
 
     if (delete_contact_from_phonebook(pb, id, 1) == 0)
     {
@@ -420,8 +420,8 @@ void menu3(phonebook *pb)
     }
 
     printf("Введите id контакта для редактирования: ");
-    unsigned long long id;
-    scanf("%llu", &id);
+    uint64_t id;
+    scanf("%lu", &id);
 
     printf("Форматные строки для редактирования:\n");
     printf("1.1 - фамилия, 1.2 - имя, 1.3 - отчество\n");
@@ -487,8 +487,8 @@ void menu5(phonebook *pb)
     }
 
     printf("Введите id контакта для добавления ссылки: ");
-    unsigned long long id;
-    scanf("%llu", &id);
+    uint64_t id;
+    scanf("%lu", &id);
 
     contact *c = find_by_id(pb, id);
 
@@ -525,8 +525,8 @@ void menu6(phonebook *pb)
     }
 
     printf("Введите id контакта для добавления почты: ");
-    unsigned long long id;
-    scanf("%llu", &id);
+    uint64_t id;
+    scanf("%lu", &id);
 
     contact *c = find_by_id(pb, id);
 
@@ -562,8 +562,8 @@ void menu7(phonebook *pb)
     }
 
     printf("Введите id контакта для добавления телефона: ");
-    unsigned long long id;
-    scanf("%llu", &id);
+    uint64_t id;
+    scanf("%lu", &id);
 
     contact *c = find_by_id(pb, id);
 
@@ -599,8 +599,8 @@ void menu8(phonebook *pb)
     }
 
     printf("Введите id контакта для удаления ссылки: ");
-    unsigned long long id;
-    scanf("%llu", &id);
+    uint64_t id;
+    scanf("%lu", &id);
 
     contact *c = find_by_id(pb, id);
 
@@ -639,8 +639,8 @@ void menu9(phonebook *pb)
     }
 
     printf("Введите id контакта для удаления почты: ");
-    unsigned long long id;
-    scanf("%llu", &id);
+    uint64_t id;
+    scanf("%lu", &id);
 
     contact *c = find_by_id(pb, id);
 
@@ -679,8 +679,8 @@ void menu10(phonebook *pb)
     }
 
     printf("Введите id контакта для удаления телефона: ");
-    unsigned long long id;
-    scanf("%llu", &id);
+    uint64_t id;
+    scanf("%lu", &id);
 
     contact *c = find_by_id(pb, id);
 
@@ -719,8 +719,8 @@ void menu11(phonebook *pb)
     }
 
     printf("Введите id контакта для поиска: ");
-    unsigned long long id;
-    scanf("%llu", &id);
+    uint64_t id;
+    scanf("%lu", &id);
 
     contact *c = find_by_id(pb, id);
 
@@ -747,7 +747,7 @@ void menu12(phonebook *pb)
     char last_name[LAST_NAME_LENGTH];
     scanf("%s", last_name);
 
-    unsigned int count;
+    uint64_t count;
     contact **cs = find_by_last_name(pb, last_name, &count);
 
     if (cs == NULL)
@@ -757,7 +757,7 @@ void menu12(phonebook *pb)
     else
     {
         printf("Найдены контакты:\n");
-        for (unsigned int i = 0; i < count; i++)
+        for (uint64_t i = 0; i < count; i++)
         {
             print_contact(cs[i]);
         }
